+++ Crafting/CreativeFilterWrapper.cs
@@ -1,11 +1,12 @@
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 using Terraria;
 using Terraria.GameContent.Creative;
 using Terraria.UI;
 
 namespace BetterInventory.Crafting;
 
-public sealed class CreativeFilterWrapper : IItemEntryFilter {
+public sealed class CreativeFilterWrapper : IRecipeFilter {
     public IItemEntryFilter Filter { get; }
     public int Index { get; }
 
@@ -16,6 +17,6 @@ public sealed class CreativeFilterWrapper : IItemEntryFilter {
     public bool FitsFilter(Item entry) => Filter.FitsFilter(entry);
     public string GetDisplayNameKey() => Filter.GetDisplayNameKey();
     public UIElement GetImage() => Filter.GetImage();
-
-    public Rectangle GetSourceFrame() => RecipeFiltering.s_recipeFilters.Frame(11, 1, frameX: Index);
+    public Texture2D GetSource(bool available) => available ? RecipeFiltering.recipeFilters.Value : RecipeFiltering.recipeFiltersGray.Value;
+    public Rectangle GetSourceFrame() => RecipeFiltering.recipeFilters.Frame(horizontalFrames: 11, frameX: Index, sizeOffsetX: -2);
 }